requirements.txt
opencv-python-headless
face_recognition
numpy
imagehas
Pillow
